+++ ChangeLog.md
@@ -8,6 +8,7 @@
   `aws:UserAgent`, `s3:x-amz-content-sha256`, `s3:x-amz-copy-source`, `s3:delimiter`,
   `s3:prefix`
 - Add default content-type for objects
+- Fix bug in parsing HTTP range requests
 
 ## [1.0.0] 2024-10-15
 - Added basic policy and authentication support